+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
+### Added
+
+- Adding error message if target element is not available in the DOM.
+
 ## [2.0.2][] - 2020-02-26
 
 ### Updated
@@ -55,7 +59,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [2.0.0]: https://github.com/willmendesneto/ngx-copy-to-clipboard/tree/v2.0.0
 [unreleased]: https://github.com/willmendesneto/ngx-copy-to-clipboard/compare/v2.0.1...HEAD
 [2.0.1]: https://github.com/willmendesneto/ngx-copy-to-clipboard/tree/v2.0.1
-
-
-[Unreleased]: https://github.com/willmendesneto/ngx-copy-to-clipboard/compare/v2.0.2...HEAD
+[unreleased]: https://github.com/willmendesneto/ngx-copy-to-clipboard/compare/v2.0.2...HEAD
 [2.0.2]: https://github.com/willmendesneto/ngx-copy-to-clipboard/tree/v2.0.2